+++ app/components/PriceDisplay/index.js
@@ -1,0 +1,1 @@
+export { default } from './PriceDisplay'
